pyproject.toml
[project]
name = "nl-opendata-mcp"
version = "0.4.2"
description = "MCP Server for Dutch government open data (CBS, data.overheid.nl)"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
    { name = "Oleksandr", email = "oleksandr.vasyliev@proton.me" }
]
keywords = ["mcp", "cbs", "opendata", "netherlands", "ai", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pandas",
    "fastmcp>=2.13.2",
    "duckdb",
    "httpx[http2]>=0.28.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.0.0",
    "seaborn>=0.13.2",
    "pytest-asyncio>=1.3.0",
]

[project.urls]
Repository = "https://github.com/soulnai/nl-opendata-mcp"
Homepage = "https://github.com/soulnai/nl-opendata-mcp"

[project.scripts]
nl-opendata-mcp = "nl_opendata_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["nl_opendata_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
